+++ mqtt/mqtt-bridge/src/persist/mod.rs
@@ -1,14 +1,14 @@
 mod loader;
 mod publication_store;
-mod storage;
 mod waking_state;
 
-use self::storage::ring_buffer::error::RingBufferError;
 use bincode::Error as BincodeError;
 pub use loader::MessageLoader;
 pub use publication_store::PublicationStore;
 use serde::{Deserialize, Serialize};
 pub use waking_state::{memory::WakingMemoryStore, StreamWakeableState};
+
+use self::waking_state::ring_buffer::error::RingBufferError;
 
 /// Keys used in persistence.
 /// Ordered by offset
@@ -26,8 +26,9 @@ pub enum PersistError {
 // This might replace the PersistError or merge with it.
 #[derive(Debug, thiserror::Error)]
 pub enum StorageError {
-    #[error("RingBuffer error occurred")]
+    #[error("RingBuffer error occurred {0}")]
     RingBuffer(#[from] RingBufferError),
-    #[error("Serialization error occurred")]
+    
+    #[error("Serialization error occurred {0}")]
     Serialization(#[from] BincodeError),
 }
