+++ mqtt/mqtt-bridge/src/persist/waking_state/ring_buffer/serialize.rs
@@ -1,27 +1,26 @@
-use bincode::{deserialize, serialize, serialized_size, Result as BincodeResult};
 use serde::{Deserialize, Serialize};
-use std::fmt::Debug;
+use bincode::Result as BincodeResult;
 
 #[allow(dead_code)]
 pub(crate) fn binary_serialize<T>(something: &T) -> BincodeResult<Vec<u8>>
 where
-    T: Serialize + Debug,
+    T: Serialize,
 {
-    serialize(something)
+    bincode::serialize(something)
 }
 
 #[allow(dead_code)]
 pub(crate) fn binary_serialize_size<T>(something: &T) -> BincodeResult<usize>
 where
-    T: Serialize + Debug,
+    T: Serialize,
 {
-    serialized_size(something).map(|x| x as usize)
+    bincode::serialized_size(something).map(|x| x as usize)
 }
 
 #[allow(dead_code)]
 pub(crate) fn binary_deserialize<'de, T>(bytes: &'de [u8]) -> BincodeResult<T>
 where
-    T: Deserialize<'de> + Debug,
+    T: Deserialize<'de>,
 {
-    deserialize::<T>(bytes)
+    bincode::deserialize::<T>(bytes)
 }
